requirements.txt
# Telegram Scraper
telethon==1.34.0
python-dotenv==1.0.0

# OCR for image-based prices
paddlepaddle>=2.6.2
paddleocr==2.7.3

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Utilities
aiohttp==3.9.1
python-dateutil==2.8.2
apscheduler==3.10.4
